requirements.txt
streamlit==1.29.0
PyPDF2==3.0.1
PyPDF4==1.27.0
pdfminer.six==20221105
pytesseract==0.3.10
opencv-python==4.8.1.78
pdf2image==1.16.3
camelot-py==0.11.0
tabula-py==2.8.2
pdfplumber==0.10.3
PyMuPDF==1.23.8
easyocr==1.7.1
pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
matplotlib==3.8.2 